feat: Introduce timeout for coroutine execution
This commit modifies the coroutine execution in `MainActivity.kt`.
The `executar()` function call within the `binding.buttonIniciar` click listener is now wrapped with `withTimeout(5000L)`. This ensures that the `executar()` suspend function will be cancelled if it does not complete within 5 seconds.

Additionally, the `delay(1000)` call within the `executar()` function has been updated to `delay(1000L)` to explicitly use a Long value for the time duration.

Signed-off-by: italosassuncao <italo18sil@outlook.com>