feat: Update Coroutine scope and UI behavior
This commit modifies the Coroutine implementation in `MainActivity.kt`.
The Coroutine scope for the main UI updates has been changed from `CoroutineScope(Dispatchers.Main)` to `MainScope()`.

Additionally, the text of `binding.buttonIniciar` is now updated within the Coroutine loop to display the current execution index and the name of the current thread.

Signed-off-by: italosassuncao <italo18sil@outlook.com>